+++ Software/wittypi/utilities.sh
@@ -148,7 +148,7 @@ if [ -z ${I2C_MC_ADDRESS+x} ]; then
 
   TIME_UNKNOWN=0
 
-  SOFTWARE_VERSION='5.57'
+  SOFTWARE_VERSION='5.58'
 
   readonly LOCAL_TZ='Europe/London'
 fi